Consolidated related functions into the same files.
