+++ js/index.js
@@ -1,17 +1,17 @@
+export * from "./WasmArray.js";
+
 export * from "./initializeSparseMatrix.js";
 export * from "./initializeSparseMatrixFromHDF5Buffer.js";
-export * from "./WasmArray.js";
+
 export * from "./computePerCellQCMetrics.js";
 export * from "./computePerCellQCFilters.js";
 export * from "./filterCells.js";
+
 export * from "./logNormCounts.js";
 export * from "./modelGeneVar.js";
 export * from "./runPCA.js";
 
-export * from "./buildNeighborSearchIndex.js";
 export * from "./findNearestNeighbors.js";
-export * from "./buildSNNGraph.js";
 export * from "./clusterSNNGraph.js";
-
 export * from "./runTSNE.js";
 export * from "./runUMAP.js";
